+++ CHANGELOG.md
@@ -7,11 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-## [1.0.1] - 17.04.2023
-
 ### Added
 
 - Added mfplat patch integration
+- Added arguments and symlinks editor to sandbox settings
 
 ### Fixed